Extract minidump creation
