+++ src/common/exception/minidump.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "../utils/nt.hpp"
+
+namespace exception
+{
+	std::string create_minidump(LPEXCEPTION_POINTERS exceptioninfo);
+}
